WIP code adds
Updated .gitignore to allow refresh_status.json which will be an internal log to use for the upcoming smart refresh functionality. This may change and will be moved to a separate DB.

Added LOGIC.md doc. This details my thought processes on some of the logic needing to be programmed in this code. Not only for public reading and commenting if necessary but also getting my thoughts into something written otherwise I WILL forget!

More test files as I learn some new python functions for this project as I go along. This is merely to migrate them between my machines and will disappear as more functional code gets written.